+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
-- Support `${classapath:path/to/eca/classpath/file}` in dynamic string parse.
-- Support default env values in `${env:MY_ENV:default-value}`.
-- Deprecate configs: 
+- Better config values dynamic string parse:
+    - Support `${classapath:path/to/eca/classpath/file}` in dynamic string parse.
+    - Support `${netrc:api.foo.com}` in dynamic string parse to parse keys.
+    - Support default env values in `${env:MY_ENV:default-value}`.
+    - Support for ECA_CONFIG and custom config file.
+- Deprecate configs:
   - `systemPromptFile` in favor of `systemPrompt` using `${file:...}` or `${classpath:...}`
   - `urlEnv` in favor of `url` using `${env:...}`
+  - `keyEnv` in favor of `key` using `${env:...}`
+  - `keyRc` in favor of `key` using `${netrc:...}`
 
 ## 0.83.0